Renamed createPackageJson to create-package-json as per convention and removed a dependency that is no longer being used
