+++ package.json
@@ -31,8 +31,7 @@
     "chalk": "^1.1.3",
     "figlet": "^1.2.0",
     "inquirer": "^1.2.3",
-    "is-valid-path": "^0.1.1",
-    "package-writer": "^0.1.1"
+    "is-valid-path": "^0.1.1"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
