+++ src/create-package-json.js
